added youtube play function
